Revert "nixcabal: remove cabal-install"
This reverts commit 776cb4d54609d4b4057f2f4090aa26f14f7304ac. While it
is possible to a Nix-based cabal project using nix-build, it will result
in more rebuilds than necessary, since the "result" symlink that's
deposited in the project causes a subsequent to build to compute a
different nix hash, even if no other files have changed. I'm following
along with the tutorials in https://github.com/Gabriel439/haskell-nix/
and they recommend using cabal for this reason.